Port TaleLib to Hytale server 0.6.0-pre API
Hytale replaced its own vector types with JOML and reworked several
subsystems. This ports TaleLib onto the current pre-release API so it
compiles against com.hypixel.hytale:Server:0.6.0-pre.12.1.

- vectors: com.hypixel.hytale.math.vector.Vector3{d,f,i} and
  com.hypixel.hytale.protocol.Vector3f -> org.joml equivalents;
  assign() -> set(), scale() -> mul(), distanceTo() -> distance()
- rotations: entity/teleport/spawn rotations now use Rotation3f, with
  its (pitch, yaw, roll) component order applied consistently
- inventory: rebuilt on InventoryComponent.getCombined/InventoryUtils
  instead of the removed Inventory.combinedEverything accessors
- npc: state and target helpers now resolve StateSupport and
  MarkedEntitySupport components directly
- worldmap: MarkerProvider now collects through MarkersCollector,
  personal markers use UserMapMarker, map images encode to a palette,
  and settings follow the new UpdateWorldMapSettings shape
- hud: CustomUIHud is keyed and registered via add/removeCustomHud
- misc: equipment updates use EquipmentUpdate, command senders resolve
  through PlayerRef, message underline takes a plain Boolean

Build moves to Java 25, Kotlin 2.4.10 and the Hytale maven repos.

+++ src/main/kotlin/com/github/ssquadteam/talelib/time/TimeExtensions.kt
@@ -1,7 +1,7 @@
 package com.github.ssquadteam.talelib.time
 
 import com.hypixel.hytale.component.Store
-import com.hypixel.hytale.math.vector.Vector3f
+import org.joml.Vector3d
 import com.hypixel.hytale.server.core.modules.time.WorldTimeResource
 import com.hypixel.hytale.server.core.universe.PlayerRef
 import com.hypixel.hytale.server.core.universe.world.World
@@ -97,7 +97,7 @@ fun World.getMoonPhase(): Int? {
     return getTimeResource()?.moonPhase
 }
 
-fun World.getSunDirection(): Vector3f? {
+fun World.getSunDirection(): Vector3d? {
     return getTimeResource()?.sunDirection
 }
